Rename "common" to "utilities" in MavenTests
