+++ tests/MavenTests/testCLI.h
@@ -4,7 +4,7 @@
 // #define _GLIBCXX_USE_CXX11_ABI 0
 
 #include <string>
-#include "common.h"
+#include "utilities.h"
 
 #include <QtTest>
