+++ tests/MavenTests/testScan.h
@@ -7,7 +7,7 @@
 #include <sstream>
 #include <fstream>
 #include <string.h>
-#include "common.h"
+#include "utilities.h"
 #include "mzSample.h"
