+++ tests/MavenTests/testbase64.h
@@ -6,7 +6,7 @@
 #include <sstream>
 #include <fstream>
 #include <string.h>
-#include "common.h"
+#include "utilities.h"
 #include "base64.h"
