+++ tests/MavenTests/utilities.cpp
@@ -1,4 +1,4 @@
-#include "common.h"
+#include "utilities.h"
 
 namespace maventests {
     Databases database;
